lesson 9: make todolist
