+++ Занятие 9/TodoList/TodoService.cs
@@ -3,35 +3,11 @@ using System;
 
 namespace TodoList
 {
-	public class TodoContext : DbContext
-	{
-		public DbSet<TodoItem> TodoItems { get; set; }
-
-		public string DbPath { get; }
-
-		public TodoContext()
-		{
-			var folder = Environment.SpecialFolder.LocalApplicationData;
-			var path = Environment.GetFolderPath(folder);
-			DbPath = System.IO.Path.Join(path, "todo.db");
-		}
-
-		public TodoContext(DbContextOptions<TodoContext> options) : base(options) {}
-
-		protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-		{
-			if (!optionsBuilder.IsConfigured) 
-			{
-				optionsBuilder.UseSqlite($"Data Source={DbPath}");
-			}
-		}
-	}
-
 	public class TodoService
 	{
-		private readonly TodoContext _context;
+		private readonly AppDbContext _context;
 
-		public TodoService(TodoContext context)
+		public TodoService(AppDbContext context)
 		{
 			_context = context;
 			_context.Database.EnsureCreated();
